+++ core/designsystem/src/main/java/com/susu/core/designsystem/component/text/SelectDateRow.kt
@@ -1,4 +1,4 @@
-package com.susu.feature.envelopeadd.content.component
+package com.susu.core.designsystem.component.text
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Row
@@ -8,11 +8,12 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
+import com.susu.core.designsystem.R
 import com.susu.core.designsystem.theme.Gray100
 import com.susu.core.designsystem.theme.Gray30
 import com.susu.core.designsystem.theme.SusuTheme
+import com.susu.core.ui.extension.susuClickable
 import com.susu.core.ui.util.AnnotatedTextByIndex
-import com.susu.feature.sent.R
 import java.time.LocalDateTime
 
 private val currentDate = LocalDateTime.now()
@@ -23,9 +24,11 @@ fun SelectDateRow(
     year: Int? = null,
     month: Int? = null,
     day: Int? = null,
+    suffix: String? = null,
+    onClick: () -> Unit = {},
 ) {
     Row(
-        modifier = modifier,
+        modifier = modifier.susuClickable(rippleEnabled = false, onClick = onClick),
         horizontalArrangement = Arrangement.spacedBy(SusuTheme.spacing.spacing_xxs),
         verticalAlignment = Alignment.Bottom,
     ) {
@@ -35,7 +38,7 @@ fun SelectDateRow(
             val dayStringRange = monthStringRange.last + 2..monthStringRange.last + 2 + currentDate.dayOfMonth.toString().length
             AnnotatedTextByIndex(
                 originalText = stringResource(
-                    R.string.sent_envelope_add_date,
+                    R.string.word_date_format,
                     currentDate.year,
                     currentDate.month.value,
                     currentDate.dayOfMonth,
@@ -51,7 +54,7 @@ fun SelectDateRow(
         } else {
             Text(
                 text = stringResource(
-                    R.string.sent_envelope_add_date,
+                    R.string.word_date_format,
                     year,
                     month,
                     day,
@@ -59,6 +62,9 @@ fun SelectDateRow(
                 style = SusuTheme.typography.title_xl,
                 color = Gray100,
             )
+        }
+        if (suffix != null) {
+            Text(text = suffix, style = SusuTheme.typography.title_l)
         }
     }
 }
