+++ feature/sent/src/main/java/com/susu/feature/envelopeadd/content/date/DateContent.kt
@@ -19,10 +19,9 @@ import com.susu.core.designsystem.component.bottomsheet.datepicker.SusuDatePicke
 import com.susu.core.designsystem.theme.Gray60
 import com.susu.core.designsystem.theme.SusuTheme
 import com.susu.core.ui.extension.collectWithLifecycle
-import com.susu.core.ui.extension.susuClickable
 import com.susu.core.ui.util.AnnotatedText
 import com.susu.core.ui.util.currentDate
-import com.susu.feature.envelopeadd.content.component.SelectDateRow
+import com.susu.core.designsystem.component.text.SelectDateRow
 import com.susu.feature.sent.R
 import java.time.LocalDateTime
 
@@ -84,10 +83,7 @@ fun DateContent(
             year = uiState.date?.year,
             month = uiState.date?.monthValue,
             day = uiState.date?.dayOfMonth,
-            modifier = Modifier.susuClickable(
-                rippleEnabled = false,
-                onClick = onClickDateText,
-            ),
+            onClick = onClickDateText,
         )
     }
